modify delete category

+++ XuongMay.Services/Service/CategoryService.cs
@@ -1,4 +1,5 @@
-﻿using System.Security.Claims;
+﻿using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
 using XuongMay.Contract.Repositories.Entity;
 using XuongMay.Contract.Repositories.Interface;
 using XuongMay.Contract.Services.Interface;
@@ -77,6 +78,12 @@ namespace XuongMay.Services.Service
             {
                 throw new BaseException.BadRequestException("Bad Request", "Cannot delete active categories");
             }
+            bool hasProducts = await _unitOfWork.GetRepository<Products>().Entities
+                                     .AnyAsync(p => p.CategoryId.Equals(categoryTemp.Id));
+            if (hasProducts)
+            {
+                throw new BaseException.BadRequestException("Bad Request", "Cannot delete active categories");
+            }
             await _unitOfWork.GetRepository<Category>().DeleteAsync(categoryTemp.Id);
             await _unitOfWork.GetRepository<Category>().SaveAsync();
         }
@@ -91,6 +98,12 @@ namespace XuongMay.Services.Service
             }
             if (categoryTemp.IsDelete)
             {
+                throw new BaseException.BadRequestException("Bad Request", "Cannot delete active categories");
+            }
+
+            bool hasProducts = await _unitOfWork.GetRepository<Products>().Entities
+                                     .AnyAsync(p => p.CategoryId.Equals(categoryTemp.Id));
+            if(hasProducts){
                 throw new BaseException.BadRequestException("Bad Request", "Cannot delete active categories");
             }
             categoryTemp.IsDelete = true;
